+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 **Internal:**
 
+- Add a metric that counts span volume in the root project for dynamic sampling (`c:spans/count_per_root_project@none`) ([#4134](https://github.com/getsentry/relay/pull/4134))
 - Emit `extracted_from_indexed` tag for the usage metrics. ([#4158](https://github.com/getsentry/relay/pull/4158))
 
 ## 24.10.0